# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Internship Recommendation Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, email, skills, domains, experienceLevel, location, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        location:
          type: string
        bio:
          type: string
        careerGoals:
          type: string
        googleId:
          type: string
        avatarUrl:
          type: string
        createdAt:
          type: string
    CreateStudentBody:
      type: object
      required: [name, email, skills, domains, experienceLevel, location]
      properties:
        name:
          type: string
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        location:
          type: string
        bio:
          type: string
        careerGoals:
          type: string
    UpdateStudentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        location:
          type: string
        bio:
          type: string
        careerGoals:
          type: string
...